crm_campaign/tealcrm/requirements.txt
#Web service
django==4.0.0
django-environ
requests
simplejson
xlwings
openpyxl
markdown
django-filter
#new
dj-database-url
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-redis
channels-redis
celery
redis
psycopg2-binary
gunicorn
pyyaml
uritemplate
imutils
requests_html
aiohttp
django-debug-toolbar
html2text
furl
django-admin-sortable2
django-crispy-forms
django-formtools
#to
